% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_raster_bin.R
\name{add_raster_bin}
\alias{add_raster_bin}
\title{Add 'bin' to dataframe with coordinates}
\usage{
add_raster_bin(
  ras,
  df,
  x = "long",
  y = "lat",
  crs_df = 4326,
  add_xy = TRUE,
  add_cell = !add_xy,
  return_old_xy = FALSE,
  add_val = FALSE,
  ...
)
}
\arguments{
\item{ras}{SpatRaster, or path to raster (needs to be path if running in
parallel), with cell numbers to extract}

\item{df}{Dataframe with x/y columns}

\item{x, y}{Character. Name of columns in \code{df} with x and y coordinates}

\item{crs_df}{Single length vector. What crs are x and y?}

\item{add_xy}{Logical. Generate (centroid) x and y coords from cell? If TRUE,
these will be returned as \code{cell_x} and \code{cell_y} where 'x' and 'y' are as
defined by arguments \code{x} and \code{y}.}

\item{add_cell}{Logical. Add the cell id to the data frame?}

\item{return_old_xy}{Logical. If true, the original x and y values will be
returned as \code{old_x} and \code{old_y} where 'x' and 'y' are as per arguments \code{x}
and \code{y}.}

\item{add_val}{Logical. If true the value(s) for cell will be extracted using
\code{terra::extract()}.The names of any columns resulting from \code{add_val} will be
the same as the names in \code{ras}.}

\item{...}{Passed to \code{envRaster::make_env_stack()}}
}
\value{
\code{df} (as tibble) with additional column(s):
\itemize{
\item \code{cell} with cell numbers from \code{ras}, if \code{add_cell}
\item \code{cell_[x]} and \code{cell_[y]}, if \code{add_xy}
\item If \code{add_val} any values from \code{ras}
Irrespective of the crs of \code{ras}, any returned x and y values will be in
\code{crs_df}.  If \code{return_old_xy} is \code{FALSE}, the original \code{x} and \code{y} columns
will be dropped.
}
}
\description{
Direct descendant of \code{envRaster::add_raster_cell()}. Adding \code{cell} is now
optional (and defaults to \code{FALSE}).
}
\examples{

  #library(envRaster)

  out_dir <- file.path(system.file(package = "envRaster"), "examples")

  base_file <- fs::path(out_dir, "cube", "base.tif")

  if(!file.exists(base_file)) {

    make_base_grid(aoi
                   , out_res = 30
                   , out_epsg = 8059
                   , out_file = base_file
                   , datatype = "INT1U"
                   , overwrite = TRUE
                   )
  }

  base <- terra::rast(base_file)

  # random sample of points within the aoi
  sample <- matrix(nrow = 2, ncol = 2) \%>\% # include some NA x and y values
    rbind(sf::st_sample(aoi, 100) \%>\%
            sf::st_transform(crs = 4326) \%>\%
            sf::st_coordinates()
          ) \%>\%
    tibble::as_tibble() \%>\%
    dplyr::mutate(attribute = sample(letters[1:10], nrow(.), replace = TRUE))

  # add just cell centroids to sample
  add_raster_bin(ras = base
                 , df = sample
                 , x = "X"
                 , y = "Y"
                 )

  # add cell and cell centroids to sample
  add_raster_bin(base
                 , sample
                 , add_cell = TRUE
                 , x = "X"
                 , y = "Y"
                 )

  # add cell, cell centroids and original 'x' and 'y' to sample
  add_raster_bin(base
                 , sample
                 , add_cell = TRUE
                 , return_old_xy = TRUE
                 , x = "X"
                 , y = "Y"
                 )

  # add cell, cell centroids, cell values and original 'x' and 'y' to sample
  add_raster_bin(base
                  , sample
                  , add_cell = TRUE
                  , return_old_xy = TRUE
                  , add_val = TRUE
                  , x = "X"
                  , y = "Y"
                  )


}
