% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aoi}
\alias{aoi}
\title{Simple feature to define a geographic area of interest.}
\format{
A simple feature with 1 rows and 2 variables:
\describe{
\item{area}{Area of simple feature in sqare metres.}
\item{geometry}{List column of geometry.}
...
}
}
\usage{
aoi
}
\description{
\code{aoi} defines an area in the northern Murray Mallee in South Australia.
}
\keyword{datasets}
