% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_index.R
\name{fix_index}
\alias{fix_index}
\title{Summarise and fix common problems with raster indices}
\usage{
fix_index(vec, sum_func = mean, rm_na = TRUE, make_int = TRUE)
}
\arguments{
\item{sum_func}{how to summarise index if more than one value is returned}

\item{rm_na}{passed to \code{sum_func} \code{na.rm} argument (which also means
\code{sum_func} needs an \code{na.rm} argument)}

\item{make_int}{Logical. If true the result is multiplied by 10000 and
an integer is returned.}
}
\value{
Numeric of length(1).
}
\description{
e.g. \code{envRaster::ndvi}. \code{Inf} or \code{-Inf} returned as \code{NA}.
}
