% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envGeomorph.R
\name{geomorph_ras}
\alias{geomorph_ras}
\title{Landform Classification after Jasiewicz and Stepinkski (2013)}
\usage{
geomorph_ras(
  ras,
  outfile = NULL,
  windowSize = 11,
  flatnessThresh = 1,
  doNew = TRUE
)
}
\arguments{
\item{ras}{input DTM (RasterLayer class).}

\item{outfile}{Name of output file. This string will have '_geomorph'
appended. Passed to writeRaster.}

\item{windowSize}{size (in terms of cells per side) of the neighborhood
(moving window) to be used; it must be an odd integer.}

\item{flatnessThresh}{If slope < flatnessThresh then flat.}

\item{doNew}{Logical. If outfile already exists, redo anyway.}
}
\value{
raster.
}
\description{
from https://github.com/m-saenger/geomorphon/blob/master/geomorphon.R
}
\details{
M. Sänger  10/2019

Raster data to be on equally spaced grid (e.g. km)
Output very sensitive to flatness threshold
Function not optmised for performance - run time very long for large data sets
}
