% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ftp_files.R
\name{get_ftp_files}
\alias{get_ftp_files}
\title{Download ftp files}
\usage{
get_ftp_files(
  paths,
  path_from = "ftp://qld.auscover.org.au",
  path_to = "../../data/raster/dynamic/AusCover",
  check_fun = NULL,
  force_new = FALSE
)
}
\arguments{
\item{paths}{Paths to download. Possibly from \code{\link[=search_ftp_files]{search_ftp_files()}}.}

\item{path_from, path_to}{Change \code{path_from} to \code{path_to} in the downloaded
file.}

\item{check_fun}{A function. Used to check if any previously downloaded
files can be read. If \code{check_fun} produces an error, those files will be
deleted before running \code{download.file}.}

\item{force_new}{Logical. If true, new files will be downloaded even if they
already exist in \code{path_to}.}
}
\value{
Called for side effect of saving ftp files to local directory
\code{path_to}.
}
\description{
Download ftp files
}
