% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_env_layers.R
\name{knit_env_layers}
\alias{knit_env_layers}
\title{Insert the environmental variables descriptions}
\usage{
knit_env_layers(ids)
}
\arguments{
\item{ids}{Character. Vector of ids (usually from env_df$blah_id) over which
to expand the environmental data sources and layers.}
}
\value{
Expanded markdown child document.
}
\description{
Describes
}
