% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envGeomorph.R
\name{land_class}
\alias{land_class}
\title{Title R function for landform classification on the basis on Topographic Position Index}
\usage{
land_class(
  ras,
  outfile = NULL,
  doNew = TRUE,
  sn = 3,
  ln = 7,
  n.classes = "six",
  win = "rectangle"
)
}
\arguments{
\item{ras}{input DTM (RasterLayer class).}

\item{outfile}{character. Optional path filename passed to writeRaster.}

\item{doNew}{logical. If FALSE, will only process x if outfile does not exist.}

\item{sn}{size (in terms of cells per side) of the neighborhood (moving
window) to be used; it must be an odd integer. s(mall) n(eighbourhood).}

\item{ln}{if the 10-class classification is selected, this paramenter sets
the l(arge) n(eighborhood) to be used.}

\item{n.classes}{"six" or "ten" for a six- or ten-class landform
classification.}

\item{win}{character. Shape of the moving window used by spatialEco::tpi.}
}
\description{
The function allows to perform landform classification on the basis of the Topographic Position Index calculated from an input Digital Terrain Model (RasterLayer class).
}
\details{
The TPI is the difference between the elevation of a given cell and the average elevation of the surrounding cells in a user defined moving window.
For landform classification, the TPI is first standardized and then thresholded; to isolate certain classes, a slope raster (which is internally worked out) is also needed.\cr
For details about the implemented classification, see: http://www.jennessent.com/downloads/tpi_documentation_online.pdf.\cr

Two methods are available:\cr
\itemize{
\item the first (devised by Weiss) produces a 6-class landform classification comprising
-- valley\cr
-- lower slope\cr
-- flat slope\cr
-- middle slope\cr
-- upper clope\cr
-- ridge\cr
-the second (devised by Jennes) produces a 10-class classification comprising
-- canyons, deeply incised streams\cr
-- midslope drainages, shallow valleys\cr
-- upland drainages, headwaters\cr
-- u-shaped valleys\cr
-- plains\cr
-- open slopes\cr
-- upper slopes, mesas\cr
-- local ridges, hills in valleys\cr
-- midslope ridges, small hills\cr
-- mountain tops, high ridges\cr
The second classification is based on two TPI that make use of two neighborhoods (moving windows) of different size: a s(mall) n(eighborhood) and a l(arge) n(eighborhood),
defined by the parameters sn and ln.\cr
}
}
\keyword{landform}
