% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_env_terra.R
\name{make_env}
\alias{make_env}
\title{Create dataframe of 'cells' and their associated env data}
\usage{
make_env(x, df, cores = 1, out_file = tempfile(), limit = TRUE)
}
\arguments{
\item{x}{Raster* object.}

\item{df}{Dataframe. Must have a column called 'cell' that corresponds to the
cell numbers in x.}

\item{cores}{Numeric. Number of cores to use in \code{\link[snowfall:snowfall-b-init]{snowfall::sfInit()}}.}

\item{out_file}{Character. Path to save outputs.}

\item{limit}{Logical. If TRUE, only cells in \code{df} will be returned. Can be
useful if cells already in \code{out_file} have been orphaned.}
}
\value{
Dataframe with column \code{cell} plus columns equal to the length of x.
}
\description{
Multicore \code{\link[raster:extract]{raster::extract()}} adapted from the Stack Exchange Network \href{https://gis.stackexchange.com/questions/253618/r-multicore-approach-to-extract-raster-values-using-spatial-points}{post}
by \href{https://gis.stackexchange.com/users/41623/thiagoveloso}{thiagoveloso}.
}
\details{
This function has been surpassed by \code{get_env_data} which uses
\code{terra::extract}.
}
