% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_env_indices.R
\name{make_env_indices}
\alias{make_env_indices}
\title{Make NDVI, SAVI, NDMI and NBR2}
\usage{
make_env_indices(
  process = c("ndvi", "savi", "msavi", "ndmi", "nbr2"),
  sum_func = "mean",
  sum_func_na_rm = TRUE,
  ...
)
}
\arguments{
\item{process}{Which functions to calculate? e.g c("ndvi", "savi")}

\item{sum_func}{Named list. If there are more than one layer over which
\code{process} are calculated, how to summarise them?}

\item{sum_func_na_rm}{\code{na.rm} argument passed to \code{sum_func}}

\item{...}{nir, red, swir1, swir2 values}
}
\value{
tibble with columns
\itemize{
\item{func}{Name of the functions applied}
\item{summarise_func}{A column for each summary function, each named for
\code{summarise_func}.}
}
}
\description{
Make NDVI, SAVI, NDMI and NBR2
}
