% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_index.R
\name{make_index}
\alias{make_index}
\title{Make and fix common problems with raster indices}
\usage{
make_index(index = "ndvi", sum_func = mean, rm_na = TRUE, ...)
}
\arguments{
\item{index}{function to calculate}

\item{sum_func}{how to summarise index if more than one value is returned}

\item{rm_na}{passed to \code{sum_func} \code{na.rm} argument (which also means
\code{sum_func} needs an \code{na.rm} argument)}

\item{...}{}
}
\value{

}
\description{
e.g. \code{envRaster::ndvi}
}
