% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_raster_index.R
\name{make_raster_index}
\alias{make_raster_index}
\title{Make raster indices}
\usage{
make_raster_index(
  X,
  func = "ndvi",
  season = NA,
  epoch = NA,
  layer = 1L,
  out_dir,
  force_new = FALSE,
  func_indexes,
  out_type = "tif",
  sum_func = "mean",
  clus_obj = NULL,
  ...
)
}
\arguments{
\item{X}{stars object}

\item{func}{index}

\item{season}{summer, winter etc}

\item{epoch}{e.g. 10-19}

\item{layer}{default 1.}

\item{out_dir}{directory to save results to}

\item{func_indexes}{NULL. or index of bands that are need by \code{func}}

\item{out_type}{tif}

\item{sum_func}{how to summarise over years in epoch}

\item{...}{}

\item{cores}{Passed to \code{parallel::makeCluster} for \code{st_apply}}
}
\value{
Side effect of saving index raster to disk
}
\description{
e.g. ndvi
}
