% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msavi.R
\name{msavi}
\alias{msavi}
\title{Calculate modified soil adjusted vegetation index (MSAVI).}
\usage{
msavi(nir, red)
}
\arguments{
\item{nir}{Numeric. Near infra-red value.}

\item{red}{Numeric. Red value.
types (see \href{https://www.usgs.gov/core-science-systems/nli/landsat/landsat-soil-adjusted-vegetation-index}{USGS.GOV}).}
}
\value{
Numeric. MSAVI value.
}
\description{
Soil adjusted vegetation index (SAVI) is used to correct Normalized
Difference Vegetation Index (NDVI) for the influence of soil brightness in
areas where vegetative cover is low. MSAVI minimizes the effect of bare soil
on the Soil Adjusted Vegetation Index (SAVI). See
\href{https://www.usgs.gov/core-science-systems/nli/landsat/landsat-soil-adjusted-vegetation-index}{USGS.GOV}.
}
