% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbr2.R
\name{nbr2}
\alias{nbr2}
\title{Calculate normalised burn ratio 2 (NBR2).}
\usage{
nbr2(swir1, swir2)
}
\arguments{
\item{swir1}{Numeric. Short-wave infra-red band 1 value.}

\item{swir2}{Numeric. Short-wave infra-red band 2 value.}
}
\value{
Numeric. NBR2 value.
}
\description{
NBR2 modifies the Normalized Burn Ratio (NBR) to highlight water sensitivity
in vegetation and may be useful in post-fire recovery studies \href{https://www.usgs.gov/core-science-systems/nli/landsat/landsat-normalized-burn-ratio-2}{USGS.GOV}.
}
