% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndmi.R
\name{ndmi}
\alias{ndmi}
\title{Calculate normalised difference moisture index (NDMI).}
\usage{
ndmi(nir, swir1)
}
\arguments{
\item{nir}{Numeric. Near infra-red value.}

\item{swir1}{Numeric. Short-wave infra-red band 1 value.}
}
\value{
Numeric. NDMI value.
}
\description{
Calculate normalised difference moisture index (NDMI).
}
