% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi.R
\name{ndvi}
\alias{ndvi}
\title{Calculate normalised difference vegetation index (NDVI).}
\usage{
ndvi(nir, red)
}
\arguments{
\item{nir}{Numeric. Near infra-red value.}

\item{red}{Numeric. Red value.}
}
\value{
Numeric. NDVI value.
}
\description{
Calculate normalised difference vegetation index (NDVI).
}
