% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_auscover.R
\name{parse_auscover}
\alias{parse_auscover}
\title{Parse an AusCover file name.}
\usage{
parse_auscover(path)
}
\arguments{
\item{path}{Character. Path(s) to AusCover raster file(s).}
}
\value{
Tibble with names
\describe{
\item{satellite}{satellite category.}
\item{instrument}{tm: thematic.}
\item{product}{re: reflective.}
\item{where}{sa: South Australia.}
\item{when}{Season start and season end dates as "m"YYYYmmYYYYmm.}
\item{process}{e.g. dja is persistent green cover.}
\item{months}{Month component of \code{when} as mmmm.}
\item{year}{Year component of \code{when} as YYYY. For \code{months} = \code{1202}, \code{year}
= \code{year + 1}.}
}
}
\description{
See for example AusCover \href{https://object-store.rc.nectar.org.au/v1/AUTH_05bca33fce34447ba7033b9305947f11/data_submission_tool_attachments/e60f5125-ed2f-47cb-99a7-c9a201e44d2f/seasonal_persistent_green_landsat_filenaming_conven_h5HG2vG.txt}{persistent green}.
filenaming convention.
}
