% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_env_tif.R
\name{parse_env_tif}
\alias{parse_env_tif}
\title{Parse meta data in raster paths used among \code{env} packages.}
\usage{
parse_env_tif(
  x,
  path_col = "path",
  cube = TRUE,
  ras_type = c("\\\\.tif", "\\\\.nc"),
  ...
)
}
\arguments{
\item{x}{Either dataframe with path(s) to parse (in \code{path_col}) or character
vector of path(s) to search.}

\item{path_col}{Character. Name of column in \code{df} with with paths.}

\item{cube}{Logical. If \code{TRUE} paths are assumed to follow
\code{raster/cube_period/env-naming/band_start-date.tif}, otherwise
\code{raster/aligned/aoi-buffer-res/env-naming.tif}. Both the base directory and
file name are used in both cases.}

\item{ras_type}{Character. e.g. "tif". More than one can be provided but if
\code{make} then the \code{ras_type[1]} is used for \code{out_file} file type.}

\item{make}{Logical. If \code{FALSE} an \code{out_file} path will be returned rather
than parsed. Assumes the appropriately names columns are supplied in \code{df}.}
}
\value{
For cube paths, tibble with columns:
\describe{
\item{source}{e.g. \code{DEA} is digital earth Austrlia}
\item{collection}{e.g. \code{ga_ls8c_ard_3} is landsat 8 collection in DEA}
\item{aoi}{Area of interest}
\item{buffer}{Any buffer around area of interest}
\item{res}{Resolution of output raster in units of rasters crs}
\item{layer}{Band or layer name}
\item{start_date}{Start date for layer}
\item{period}{Period represented within layer using ISO8601-compliant time
period for regular data cubes. e.g., "P16D" for 16 days}
\item{type}{File type of \code{path}.}
\item{path}{Full (relative) path including \code{ras_type}.}
}
For static paths, tibble with columns:
\describe{
\item{season}{Season into which summary was made.}
\item{epoch}{Date range (years) into which summary was made.}
}

If \code{make}, \code{df} with additional column \code{out_file}
}
\description{
Parse meta data in raster paths used among \code{env} packages.
}
