% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_auscover.R
\name{prep_auscover}
\alias{prep_auscover}
\title{Generate lists of paths in preparation for \code{\link[=summarise_rast_paths]{summarise_rast_paths()}}.}
\usage{
prep_auscover(
  dir_local = "../../data/raster/dynamic/AusCover/landsat/",
  epoch_step = 10,
  epoch_overlap = FALSE,
  epoch_which = -1,
  filt_n = 9
)
}
\arguments{
\item{dir_local}{Character. Path to search for rasters to summarise.}

\item{epoch_step}{Numeric. How many years in an epoch?}

\item{epoch_overlap}{Logical. Should epochs overlap by one year? i.e.
\code{epoch_overlap = TRUE} gives, say, 2000-2010 and 2010-2020 whereas
\code{epoch_overalp = FALSE} gives, say, 2000-2009 and 2010-2019.}

\item{epoch_which}{Numeric. Which epochs to summarise? Zero for current
epoch, -1 for immediate previous epoch etc. -2:-1 for the last two full
epochs.}

\item{filt_n}{Numeric. Filter prepared stacks with less than \code{filt_n}
rasters.}
}
\value{
Dataframe with columns
\describe{
\item{process}{Raster product in \code{data}. See \link[envEcosystems:env]{envEcosystems::env}.}
\item{epoch}{Last two digits of initial year and final year for rasters
in \code{data}. e.g. 10-19.}
\item{season}{Season for rasters in \code{data}.}
\item{data}{List column of full path of raster files to summarise.}
}
}
\description{
Using the file naming conventions in the \href{http://data.auscover.org.au/xwiki/bin/view/Field+Sites/AusCover+Filenaming+Convention}{AusCover data}
, collate paths of previously downloaded raster files in preparation for
running \code{\link[envImport:summarise_rast_paths]{envImport::summarise_rast_paths()}}.
}
