% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ras_collection}
\alias{ras_collection}
\title{Lookup for collections of raster data}
\format{
A data frame with 10 rows and 4 variables:
\describe{
\item{source}{Short name of \code{source}}
\item{type}{What sort of data is provided in the \code{collection}?}
\item{collection}{Name of the \code{collection} within \code{source}}
\item{coll_name}{Full name of the \code{collection} within \code{source}}
\item{collection_url}{URL for \code{collection}}
\item{collection_notes}{Notes about \code{collection}}
...
}
}
\usage{
ras_collection
}
\description{
Lookup for collections of raster data
}
\keyword{datasets}
