% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ras_source}
\alias{ras_source}
\title{Lookup for sources of raster data}
\format{
A data frame with 4 rows and 4 variables:
\describe{
\item{source}{Short name of \code{source}}
\item{source_name}{Full name of \code{source}}
\item{source_url}{URL for \code{source}}
\item{source_notes}{Notes about \code{source}}
...
}
}
\usage{
ras_source
}
\description{
Lookup for sources of raster data
}
\keyword{datasets}
