% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savi.R
\name{savi}
\alias{savi}
\title{Calculate soil adjusted vegetation index (SAVI).}
\usage{
savi(nir, red, L = 0.5)
}
\arguments{
\item{nir}{Numeric. Near infra-red value.}

\item{red}{Numeric. Red value.}

\item{L}{Numeric. Soil brightness correction factor. 0.5 for most land cover
types (see \href{https://www.usgs.gov/core-science-systems/nli/landsat/landsat-soil-adjusted-vegetation-index}{USGS.GOV}).}
}
\value{
Numeric. SAVI value.
}
\description{
SAVI is used to correct Normalized Difference Vegetation Index (NDVI) for the
influence of soil brightness in areas where vegetative cover is low
\href{https://www.usgs.gov/core-science-systems/nli/landsat/landsat-soil-adjusted-vegetation-index}{USGS.GOV}.
}
