% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ftp_files.R
\name{search_ftp_files}
\alias{search_ftp_files}
\title{Search for files on an ftp site}
\usage{
search_ftp_files(
  dir_ftp = "ftp://qld.auscover.org.au/landsat",
  dir_regex = NULL,
  recurse = FALSE
)
}
\arguments{
\item{dir_ftp}{Character. URL (directory on ftp site to search).}

\item{dir_regex}{Character. Search string to limit results from within \code{dir_ftp}.}

\item{recurse}{Logical. Recursively search \code{dir_ftp}?}
}
\value{
Tibble with columns
\describe{
\item{file}{If path is a file, this will contain \code{basename(path)}.}
\item{path}{Full path to file for use in \code{\link[=get_ftp_files]{get_ftp_files()}}.}
}
}
\description{
Defaults to \href{http://qld.auscover.org.au/public/html/index.html}{TERN}
AusCover landsat.
}
\examples{
search_ftp_files()
}
