% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_env.R
\name{summarise_env}
\alias{summarise_env}
\title{Environmental summary statistics}
\usage{
summarise_env(env_df, context, luenv_df = NULL, trans_col = "transform")
}
\arguments{
\item{env_df}{Data frame of environmental values.}

\item{context}{Character name of column(s) in \code{env_df} that define the context.}

\item{luenv_df}{Data frame with information about each layer.}

\item{trans_col}{Character. Name of column in \code{luenv_df} containing
information on how to convert raster values to the scale of \code{units}.}
}
\value{
Data frame of summary values for each layer, particularly \code{units} and
\code{transform} to convert raster layer values to \code{units}.
}
\description{
Given a data frame of environmental values, calculate summary statistics
}
