% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_intersection.R
\name{test_intersection}
\alias{test_intersection}
\title{Test for intersection of two rasters}
\usage{
test_intersection(a, b)
}
\arguments{
\item{a}{1st raster}

\item{b}{2nd raster}
}
\value{
Logical
}
\description{
Test for intersection of two rasters
}
