% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ras_xn.R
\name{test_ras_xn}
\alias{test_ras_xn}
\title{Test for intersection of two rasters}
\usage{
test_ras_xn(a, b)
}
\arguments{
\item{a, b}{raster as SpatRaster or path to raster}
}
\value{
Logical
}
\description{
Actually tests for the intersection of the extent of each raster
}
\examples{
  #library(envRaster)

  out_dir <- file.path(system.file(package = "envRaster"), "examples")

  base_file <- fs::path(out_dir, "cube", "base.tif")

  if(!file.exists(base_file)) {

    make_base_grid(aoi
                   , out_res = 30
                   , out_epsg = 8059
                   , out_file = base_file
                   , datatype = "INT1U"
                   , overwrite = TRUE
                   )
  }

  base <- terra::rast(base_file)

  test_ras_xn(base, base)

  test_ras_xn(base
              , system.file("ex/meuse.tif", package="terra")
              )
}
