% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widen_env_data.R
\name{widen_env_data}
\alias{widen_env_data}
\title{Widen env data, especially AusCover data from \code{get_env_data}.}
\usage{
widen_env_data(
  env_data_long,
  orig_df,
  context,
  env_groups = c("process", "layer", "method", "season"),
  epoch_step = 10,
  min_years = 3
)
}
\arguments{
\item{env_data_long}{Dataframe, in long format. Usually resulting from
\code{get_env_data}.}

\item{orig_df}{Dataframe containing the locations passed to \code{get_env_data}.}

\item{context}{Character. Columns in \code{orig_df} defining the context.}

\item{env_groups}{Character. Name of columns in \code{env_data_long} that define
the context of the raster. Default to 'process', 'layer', 'method' and
'season'. See \link[envEcosystems:env]{envEcosystems::env} for definitions.}

\item{epoch_step}{Numeric. If summarising several (years) rasters to one for
the given \code{context}, how many years to summarise?}

\item{min_years}{Numeric. What is the minimum acceptable number of years with
data. Points with less than \code{min_years} of data will be filtered.}
}
\value{
Site (context) by environmental data with one row of env data for
each context.
}
\description{
Output is a context by environmental variable data frame.
}
