% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_add_info.R
\name{env_add_info}
\alias{env_add_info}
\title{Add information from \code{ras_layers} to a dataframe}
\usage{
env_add_info(df, col = "name")
}
\arguments{
\item{df}{Dataframe with column \code{col}}

\item{col}{Character. Name of column in \code{col} containing information to match
to \code{ras_layers}. This is usually the \code{name} column resulting from
\code{envRaster::name_env_tif()} with argument \code{make_name} set to \code{TRUE}.}
}
\value{
tibble with extra columns from \code{envRaster::ras_layers}
}
\description{
Add information from \code{ras_layers} to a dataframe
}
