% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_env_stack.R
\name{make_env_stack}
\alias{make_env_stack}
\title{Make a spatRaster from a vector of paths}
\usage{
make_env_stack(
  predictors,
  is_env_pred = TRUE,
  aoi = NULL,
  cat_pred_levels = NULL
)
}
\arguments{
\item{predictors}{Character. Paths to rasters.}

\item{is_env_pred}{Logical. Does the naming follow that described by
\code{envRaster::name_env_tif()}}

\item{aoi}{sf. Used in \code{terra::window()}}

\item{cat_pred_levels}{Named list. Usually \code{prep$cat_pred_levels()} resulting
from \code{envSDM::prep_sdm()}}
}
\value{
spatRaster
}
\description{
Allows for some degree of matching extent; settings a window (using the aoi);
and recoding categorical raster(s) to no more than cat_pred_levels.
}
