% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_env_tif.R
\name{name_env_tif}
\alias{name_env_tif}
\title{Use meta data to name raster cube paths/files}
\usage{
name_env_tif(
  x,
  dir_only = FALSE,
  prefixes = c("sat", "use"),
  file_type = c("\\\\.tif", "\\\\.nc"),
  parse = FALSE,
  fill_null = FALSE,
  x_null = 3,
  ...
)
}
\arguments{
\item{x}{Either dataframe with path(s) to parse (in column \code{path}), character
vector of path(s) to search, or named list object.}

\item{dir_only}{Logical.}

\item{prefixes}{Character. If x is a named list, any name prefixes to remove
before matching to naming structure.}

\item{file_type}{Character. e.g. "tif". More than one can be provided but if
\code{parse} then \code{file_type[1]} is used for \code{out_file} file type.}

\item{parse}{Logical. If \code{FALSE} an \code{out_file} path will be returned rather
than parsed. Assumes the appropriately names columns are supplied in \code{df}.}

\item{fill_null}{Logical. If \code{TRUE}, will fill up to \code{x_null} definitions
with NULL (and issue a warning).}

\item{x_null}{Numeric. Even if \code{fill_null} is \code{TRUE}, if there are more than
\code{x_null} missing definitions, an error will be thrown.}
}
\value{
If \code{!parse}, tibble with columns:
\describe{
\item{vector}{Vector layer originally used to define area of interest}
\item{filt_col}{Column name from vector to filter to define area of interest}
\item{level}{Level(s) of filt_col originally filtered to define area of interest}
\item{buffer}{Any buffer around area of interest}
\item{period}{Period represented within layer using ISO8601-compliant time
period for regular data cubes. e.g., "P16D" for 16 days}
\item{res}{Resolution of output raster in units of rasters crs}
\item{source}{e.g. \code{DEA} is digital earth Austrlia}
\item{collection}{e.g. \code{ga_ls8c_ard_3} is landsat 8 collection in DEA}
\item{layer}{Band or layer name}
\item{start_date}{Start date for layer}
\item{type}{File type of \code{path}.}
\item{path}{Full (relative) path including \code{file_type}.}
}

If \code{!parse}, \code{df} with additional column \code{out_file}
}
\description{
Either create file path for saving a cube or parse the meta data from a cube
}
