% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_env_tif.R
\name{parse_env_tif}
\alias{parse_env_tif}
\title{Parse tif saved using raster file naming convention among \code{env} packages.}
\usage{
parse_env_tif(path, ...)
}
\arguments{
\item{path}{Character. Path to search for rasters to parse.}

\item{...}{Passed to \code{\link[fs:dir_ls]{fs::dir_info()}}.}
}
\value{
Dataframe with columns
\describe{
\item{process}{e.g. dja is persistent green cover.}
\item{layer}{band or layer that was summarised.}
\item{method}{Method that was used to summarise.}
\item{season}{Season into which summary was made.}
\item{epoch}{Date range (years) into which summary was made.}
\item{type}{File type of \code{path}.}
\item{path}{Full (relative) path including \code{tif}.}
}
}
\description{
Parse tif saved using raster file naming convention among \code{env} packages.
}
