% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ras_layers}
\alias{ras_layers}
\title{Lookup for raster layers}
\format{
A data frame with 93 rows and 12 variables:
\describe{
\item{source}{Short name of \code{source}}
\item{collection}{Name of the \code{collection} within \code{source}}
\item{type}{Type of layer. e.g. reflectance, indice, variability, derived}
\item{layer}{A short name}
\item{method}{Method used to summarise if more than one layer was available}
\item{units}{Units of layer}
\item{mult}{value * mult = data on original scale. Essentially irrelvant to
use of the layers though}
\item{description}{A slightly longer description of the layer than \code{layer}}
\item{indicates}{Why is this layer ecologically useful?}
\item{notes}{}
\item{method_rf}{A reference for the method used to generate the output}
\item{data_ref}{A reference for the method used to generate input layer(s)}
...
}
}
\usage{
ras_layers
}
\description{
Lookup for raster layers
}
\keyword{datasets}
