% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ras_layers}
\alias{ras_layers}
\title{Lookup for raster layers}
\format{
A data frame with 32 rows and 12 variables:
\describe{
\item{source}{Short name of \code{source}}
\item{collection}{Name of the \code{collection} within \code{source}}
\item{type}{}
\item{layer}{}
\item{method}{}
\item{season}{}
\item{year_min}{}
\item{year_max}{}
\item{units}{}
\item{description}{}
\item{indicates}{}
\item{notes}{}
\item{ref}{}
\item{data_ref}{}
...
}
}
\usage{
ras_layers
}
\description{
Lookup for raster layers
}
\keyword{datasets}
